app/requirements.txt
flask==3.0.0
flask-cors==4.0.0
requests==2.31.0
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.4
matplotlib==3.7.2
seaborn==0.12.2
lifelines==0.27.7
scipy==1.11.1
pyyaml==6.0.1
